# custom/mnzl/receivables/openapi/receivables-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MNZL purchased receivables",
    "version": "1"
  },
  "servers": [
    {
      "url": "/fineract-provider/api/v1/mnzl/receivables"
    }
  ],
  "security": [
    {
      "fineractAuthentication": []
    }
  ],
  "paths": {
    "/accounts": {
      "get": {
        "parameters": [
          {
            "name": "Fineract-Platform-TenantId",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 100
            }
          },
          {
            "name": "asOfDate",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/financialDate"
            }
          },
          {
            "name": "periodStartDate",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/financialDate"
            }
          },
          {
            "name": "eventWatermark",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/version"
            }
          },
          {
            "name": "financierOrganizationId",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Validated result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/accountsPage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "403": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "404": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "409": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "422": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "503": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          }
        }
      }
    },
    "/journals": {
      "get": {
        "parameters": [
          {
            "name": "Fineract-Platform-TenantId",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "operationIds",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/id"
              },
              "minItems": 1
            }
          },
          {
            "name": "eventWatermark",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Validated result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/nativeJournalReadback"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "403": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "404": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "409": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "422": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "503": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          }
        }
      }
    },
    "/hel-terms/{externalId}": {
      "get": {
        "parameters": [
          {
            "name": "Fineract-Platform-TenantId",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "externalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Validated result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/helTerms"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "403": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "404": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "409": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "422": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "503": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          }
        }
      }
    },
    "/accounts/{externalId}/measurement": {
      "get": {
        "parameters": [
          {
            "name": "Fineract-Platform-TenantId",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "externalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "businessDate",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/financialDate"
            }
          },
          {
            "name": "boundarySide",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/boundarySide"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Validated result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/accountMeasurement"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "403": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "404": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "409": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "422": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "503": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          }
        }
      }
    },
    "/hel-journals": {
      "get": {
        "parameters": [
          {
            "name": "Fineract-Platform-TenantId",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "loanId",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/id"
            }
          },
          {
            "name": "transactionIds",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/id"
              },
              "minItems": 1
            }
          },
          {
            "name": "eventWatermark",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Validated result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/helJournals"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "403": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "404": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "409": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "422": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "503": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          }
        }
      }
    },
...
    "/hel-funding/allocations/{operationId}": {
      "get": {
        "parameters": [
          {
            "name": "Fineract-Platform-TenantId",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "operationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Validated result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/helAdvanceAllocationOperation"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "403": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "404": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "409": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "422": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "503": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          }
        }
      }
    },
...
    "/hel-funding/operations/{operationId}": {
      "get": {
        "parameters": [
          {
            "name": "Fineract-Platform-TenantId",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "operationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Validated result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/helFundingResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "403": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "404": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "409": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "422": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "503": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          }
        }
      }
    },
...
    "/capabilities": {
      "get": {
        "parameters": [
          {
            "name": "Fineract-Platform-TenantId",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Validated result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/capabilities"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "403": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "404": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "409": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "422": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "503": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          }
        }
      }
    },
...
    "/operations/{operationId}": {
      "get": {
        "parameters": [
          {
            "name": "Fineract-Platform-TenantId",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "operationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Validated result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/financialOperation"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "403": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "404": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "409": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "422": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "503": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          }
        }
      }
    },
    "/accounts/{externalId}": {
      "get": {
        "parameters": [
          {
            "name": "Fineract-Platform-TenantId",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "externalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Validated result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/accountRead"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "403": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "404": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "409": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "422": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          },
          "503": {
            "description": "Invalid input, denied scope, missing object, version/state conflict, financial rejection or retryable infrastructure failure, respectively."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "fineractAuthentication": {
        "type": "http",
        "scheme": "basic"
      }
    },
    "schemas": {
      "boundarySide": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/boundarySide",
        "type": "string",
        "enum": [
          "BEFORE_EVENTS",
          "AFTER_EVENTS"
        ]
      },
...
      "financialDate": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/financialDate",
        "type": "string",
        "format": "date",
        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
      },
...
      "financierScope": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/financierScope",
        "type": "object",
        "properties": {
          "platformId": {
            "$ref": "#/components/schemas/id"
          },
          "financierOrganizationId": {
            "$ref": "#/components/schemas/id"
          },
          "environment": {
            "type": "string",
            "enum": [
              "local",
              "test",
              "staging",
              "production"
            ]
          },
          "ledgerEpoch": {
            "$ref": "#/components/schemas/id"
          }
        },
        "required": [
          "platformId",
          "financierOrganizationId",
          "environment",
          "ledgerEpoch"
        ],
        "additionalProperties": false
      },
      "id": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/id",
        "type": "string",
        "minLength": 1,
        "maxLength": 200
      },
...
      "version": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/version",
        "type": "string",
        "pattern": "^(?:0|[1-9][0-9]*)$"
      },
...
      "accountMeasurement": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/accountMeasurement",
        "type": "object",
        "properties": {
          "position": {
            "$ref": "#/components/schemas/position"
          },
          "measurementLegs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/measurementLeg"
            }
          },
          "eventWatermark": {
            "$ref": "#/components/schemas/version"
          }
        },
        "required": [
          "position",
          "measurementLegs",
          "eventWatermark"
        ],
        "additionalProperties": false
      },
...
      "accountsPage": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "#/components/schemas/accountsPage",
        "type": "object",
        "properties": {
          "scope": {
            "$ref": "#/components/schemas/financierScope"
          },
          "productCode": {
            "type": "string",
            "const": "MNZL_PURCHASED_RECEIVABLE"
          },
          "asOfDate": {
            "$ref": "#/components/schemas/financialDate"
          },
          "eventWatermark": {
            "$ref": "#/components/schemas/version"
          },
          "includesClosedPeriodMovements": {
            "type": "boolean",
            "const": true
          },
          "items": {
            "maxItems": 200,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/accountSummary"
            }
          },
          "nextCursor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/id"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "scope",
          "productCode",
          "asOfDate",
          "eventWatermark",
          "includesClosedPeriodMovements",
          "items",
          "nextCursor"
        ],
        "additionalProperties": false
      },
...
    }
  }
}